Share provider playback types
