+++ packages/shared/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@cliparr/shared",
+  "private": true,
+  "version": "0.4.0",
+  "type": "module",
+  "types": "./src/providers.ts",
+  "exports": {
+    "./providers": {
+      "types": "./src/providers.ts",
+      "default": "./src/providers.ts"
+    }
+  }
+}
